Add convention plugin for Git hooks setup
